Add Localization and Menu. Rebuild project

+++ src/BW/BlogBundle/Controller/Admin/PostController.php
@@ -5,8 +5,8 @@ namespace BW\BlogBundle\Controller\Admin;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 
 use BW\MainBundle\Controller\BWController;
-use BW\BlogBundle\Form\PostType;
 use BW\BlogBundle\Entity\Post;
+use BW\BlogBundle\Form\PostType;
 
 class PostController extends BWController
 {
@@ -33,43 +33,47 @@ class PostController extends BWController
         return $this->render('BWBlogBundle:Admin/Post:posts.html.twig', $data->toArray());
     }
     
-    public function postAction() {
-        
-    }
+//    public function addAction() {
+//        $data = $this->getPropertyOverload();
+//        $request = $this->get('request');
+//        
+//        $post = new Post;
+//        $form = $this->createForm(new \BW\BlogBundle\Form\PostAddType(), $post);
+//        
+//        if ($request->isMethod('POST')) {
+//            $form->handleRequest($request);
+//            
+//            if ($form->isValid()) {
+//                $em = $this->getDoctrine()->getManager();
+//                $em->persist($post);
+//                $em->flush();
+//                
+//                if ( $form->get('saveAndExit')->isClicked() ) {
+//                    return $this->redirect( $this->generateUrl('admin_posts') );
+//                }
+//                
+//                return $this->redirect( $this->generateUrl('admin_edit_post', array('id' => $post->getId())) );
+//            }
+//        }
+//        
+//        $data->form = $form->createView();
+//        return $this->render('BWBlogBundle:Admin/Post:add-post.html.twig', $data->toArray());
+//    }
     
-    public function addAction() {
+    public function postAction($id = NULL) {
         $data = $this->getPropertyOverload();
         $request = $this->get('request');
         
-        $post = new Post;
-        $form = $this->createForm(new \BW\BlogBundle\Form\PostAddType(), $post);
-        
-        if ($request->isMethod('POST')) {
-            $form->handleRequest($request);
-            
-            if ($form->isValid()) {
-                $em = $this->getDoctrine()->getManager();
-                $em->persist($post);
-                $em->flush();
-                
-                if ( $form->get('saveAndExit')->isClicked() ) {
-                    return $this->redirect( $this->generateUrl('admin_posts') );
-                }
-                
-                return $this->redirect( $this->generateUrl('admin_edit_post', array('id' => $post->getId())) );
-            }
+        if ($id) {
+            $post = $this->getDoctrine()->getRepository('BWBlogBundle:Post')->find($id);
+        } else {
+            $post = new Post;
         }
         
-        $data->form = $form->createView();
-        return $this->render('BWBlogBundle:Admin/Post:add-post.html.twig', $data->toArray());
-    }
-    
-    public function editAction($id) {
-        $data = $this->getPropertyOverload();
-        $request = $this->get('request');
-        
-        $post = $this->getDoctrine()->getRepository('BWBlogBundle:Post')->find($id);
-        $form = $this->createForm(new \BW\BlogBundle\Form\PostEditType(), $post);
+        $form = $this->createForm(new PostType(), $post);
+        if ( ! $id) {
+            $form->remove('delete');
+        }
         
         if ($request->isMethod('POST')) {
             $form->handleRequest($request);
@@ -77,15 +81,26 @@ class PostController extends BWController
             if ($form->isValid()) {
                 $em = $this->getDoctrine()->getManager();
                 
-                if ( $form->get('delete')->isClicked() ) {
-                    $em->remove($post);
-                    $em->flush();
-                    
-                    return $this->redirect( $this->generateUrl('admin_posts') );
+                if ($id) {
+                    if ( $form->get('delete')->isClicked() ) {
+                        $em->remove($post);
+                        $em->flush();
+                        
+                        $this->get('session')->getFlashBag()->add(
+                            'danger',
+                            'Статья успешно удалена из БД'
+                        );
+
+                        return $this->redirect( $this->generateUrl('admin_posts') );
+                    }
                 }
                 
                 $em->persist($post);
                 $em->flush();
+                $this->get('session')->getFlashBag()->add(
+                        'success',
+                        'Статья успешно сохранена в БД'
+                    );
                 
                 if ( $form->get('saveAndExit')->isClicked() ) {
                     return $this->redirect( $this->generateUrl('admin_posts') );
@@ -97,7 +112,12 @@ class PostController extends BWController
         
         $data->post = $post;
         $data->form = $form->createView();
-        return $this->render('BWBlogBundle:Admin/Post:edit-post.html.twig', $data->toArray());
+        
+        if ($id) {
+            return $this->render('BWBlogBundle:Admin/Post:edit-post.html.twig', $data->toArray());
+        }
+        
+        return $this->render('BWBlogBundle:Admin/Post:add-post.html.twig', $data->toArray());
     }
     
     public function deleteAction() {
